Add files via upload
The School Management System (SMS) is a system that will provide you with a bird's-eye view of the functioning of the entire educational institution. It is a management information system that helps to manage the different processes in areas like general administration, staff management, student management, grade management, course management, and class management. The information is made using the latest technologies and helps make decision-making a lot faster, more effective, and easier than ever before. It also contributes to the overall quality of education in schools.